requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-telegram-bot==20.8
sqlalchemy==2.0.25
apscheduler==3.10.4
pydantic==2.5.3
jinja2==3.1.3
python-dotenv==1.0.0
